+++ packages/pwa-kit-react-sdk/CHANGELOG.md
@@ -1,3 +1,6 @@
+## ???
+- useBlock function and Skeleton: [#2247](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2247)
+
 ## v3.9.0-dev (Oct 29, 2024)
 - Fix the performance logging util to not round duration. [#2199](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2199)
 - Add RedirectWithStatus component, allowing finer grained control of rediriects and their status code [#2173](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2173)